fix(node): blocks requests are not parallelized (#234)

+++ src/node/config/settings.go
@@ -9,5 +9,5 @@ type Settings struct {
 	MinimalTransactionFee            uint64
 	SynchronizationIntervalInSeconds int
 	ValidationIntervalInSeconds      int64
-	VerificationsCountPerValidation  int
+	VerificationsCountPerValidation  int64
 }
